requirements.txt
-i https://pypi.org/simple
amqp==5.2.0; python_version >= '3.6'
annotated-types==0.6.0; python_version >= '3.8'
asgiref==3.7.2; python_version >= '3.7'
attrs==23.2.0; python_version >= '3.7'
billiard==4.2.0; python_version >= '3.7'
boto3==1.34.42; python_version >= '3.8'
botocore==1.34.61; python_version >= '3.8'
celery[redis]==5.3.6; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.0; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
django==5.0.2; python_version >= '3.10'
djangorestframework==3.14.0; python_version >= '3.6'
drf-spectacular==0.27.1; python_version >= '3.6'
gunicorn==21.2.0; python_version >= '3.5'
inflection==0.5.1; python_version >= '3.5'
jmespath==1.0.1; python_version >= '3.7'
jsonschema==4.21.1; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
kombu==5.3.5; python_version >= '3.8'
mysqlclient==2.0.0; python_version >= '3.5'
packaging==24.0; python_version >= '3.7'
prompt-toolkit==3.0.43; python_full_version >= '3.7.0'
pydantic==2.6.4; python_version >= '3.8'
pydantic-core==2.16.3; python_version >= '3.8'
pydantic-settings==2.2.1; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
redis==5.0.3
referencing==0.33.0; python_version >= '3.8'
rpds-py==0.18.0; python_version >= '3.8'
s3transfer==0.10.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.10.0; python_version >= '3.8'
tzdata==2024.1; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.0.7; python_version >= '3.10'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13
